requirements.txt
discord==2.3.2
discord-interactions==0.4.0
discord-py-slash-command==4.2.1
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
cffi==1.16.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mysql-connector-python==8.4.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
yarl==1.9.4
